Texture problems...

+++ C-Craft++/World/Generation/Biome/Biome.cpp
@@ -15,7 +15,7 @@ ChunkBlock Biome::getBeachBlock(Rand& rand)const
 
 int Biome::getHeight(int x, int z, int chunkX, int chunkZ) const
 {
-	return m_heightGenerator.getHeight(x, z, chunkX, chunkZ);
+	return (int)m_heightGenerator.getHeight(x, z, chunkX, chunkZ);
 }
 
 int Biome::getTreeFrequency() const noexcept
